+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.58.0](https://github.com/KL-Engineering/kidsloop-hub-frontend/branches/compare/v2.58.0%0Dv2.57.2) (2022-05-18)
+
+
+### ✨ Features
+
+* **spr:** add skill perfomance spr and refinement (#113) ([5bc2217](https://github.com/KL-Engineering/kidsloop-hub-frontend/commits/5bc2217b4fb7e8e0c6e8db7b733bc2daf175f39c))
+
+### [2.57.2](https://github.com/KL-Engineering/kidsloop-hub-frontend/branches/compare/v2.57.2%0Dv2.57.1) (2022-05-18)
+
+
+### 🗑 Reverts
+
+* Revert "feat: bulk status action for users (#112)" (#117) ([9538451](https://github.com/KL-Engineering/kidsloop-hub-frontend/commits/9538451b32aa7c8e8d7d346dd1e5051386b9ac24))
+
+### [2.57.1](https://github.com/KL-Engineering/kidsloop-hub-frontend/branches/compare/v2.57.1%0Dv2.57.0) (2022-05-18)
+
+
+### 🐛 Bug Fixes
+
+* **css:** page width reverted (#115) ([fe4d38b](https://github.com/KL-Engineering/kidsloop-hub-frontend/commits/fe4d38b402700db665cb1d638b8114a86230d4d3))
+
+
+### ♻️ Chores
+
+* **DT-847:** Student Progress Report Translations (#116) ([38f5ba3](https://github.com/KL-Engineering/kidsloop-hub-frontend/commits/38f5ba30a0b755387bb6bf1ca2aa183e1142b1d2)), closes [DT-847](https://calmisland.atlassian.net/browse/DT-847)
+
+## [2.57.0](https://github.com/KL-Engineering/kidsloop-hub-frontend/branches/compare/v2.57.0%0Dv2.56.1) (2022-05-18)
+
+
+### ✨ Features
+
+* bulk status action for users (#112) ([a0a4d10](https://github.com/KL-Engineering/kidsloop-hub-frontend/commits/a0a4d10040fd71634b815e203499c689a40f9457))
+
 ### [2.56.1](https://github.com/KL-Engineering/kidsloop-hub-frontend/branches/compare/v2.56.1%0Dv2.56.0) (2022-05-13)